feat: end demo

+++ 01fundamentals/05StringsStreamsDemos/05StdStringFindMultipleOccurrences.cpp
@@ -1,14 +1,17 @@
 #include <iostream>
-#include <string>
+// #include <string>
+
 using namespace std;
+
 int main() {
   string str = "canary";
 
-  int foundIndex = str.find("a");
+  size_t foundIndex = str.find("a");
 
   while (foundIndex != string::npos) {
     cout << "a"
          << " found at " << foundIndex << " in " << str << endl;
+
     foundIndex = str.find("a", foundIndex + 1);
   }
